+++ mols2grid/templates/js/selection_actions.js
@@ -1,59 +1,74 @@
 // check all
 $('#btn-chkbox-all').click(function (e) {
-    var current_page = parseInt($("li.page-item.active > a").text());
-    var n_items = {{ n_items_per_page }};
-    var last_page = parseInt($("li.page-item > a").last().text());
     var _id = [];
     var _smiles = [];
-    for (let i = 0; i < last_page; i++) {
-        listObj.show(i * n_items + 1, n_items);
-        $("input:checkbox").each(function() {
-            this.checked = true;
-            _id.push(parseInt($(this).closest(".cell").attr("data-mols2grid-id")));
-            _smiles.push($($(this).siblings(".data-{{ smiles_col }}")[0]).text());
-        });
-    }
-    listObj.show((current_page - 1) * n_items + 1, n_items);
+    listObj.items.forEach(function (item) {
+        if (item.elm) {
+            item.elm.getElementsByTagName("input")[0].checked = true;
+        } else {
+            item.show()
+            item.elm.getElementsByTagName("input")[0].checked = true;
+            item.hide()
+        }
+        _id.push(item.values()["mols2grid-id"]);
+        _smiles.push(item.values()["data-{{ smiles_col }}"]);
+    });
+    add_selection({{ grid_id | tojson }}, _id, _smiles);
+});
+// check matching
+$('#btn-chkbox-match').click(function (e) {
+    var _id = [];
+    var _smiles = [];
+    listObj.matchingItems.forEach(function (item) {
+        if (item.elm) {
+            item.elm.getElementsByTagName("input")[0].checked = true;
+        } else {
+            item.show()
+            item.elm.getElementsByTagName("input")[0].checked = true;
+            item.hide()
+        }
+        _id.push(item.values()["mols2grid-id"]);
+        _smiles.push(item.values()["data-{{ smiles_col }}"]);
+    });
     add_selection({{ grid_id | tojson }}, _id, _smiles);
 });
 // uncheck all
 $('#btn-chkbox-none').click(function (e) {
-    var current_page = parseInt($("li.page-item.active > a").text());
-    var n_items = {{ n_items_per_page }};
-    var last_page = parseInt($("li.page-item > a").last().text());
     var _id = [];
-    for (let i = 0; i < last_page; i++) {
-        listObj.show(i * n_items + 1, n_items);
-        $("input:checkbox:checked").each(function() {
-            this.checked = false;
-            _id.push(parseInt($(this).closest(".cell").attr("data-mols2grid-id")));
-        });
-    }
-    listObj.show((current_page - 1) * n_items + 1, n_items);
+    listObj.items.forEach(function (item) {
+        if (item.elm) {
+            item.elm.getElementsByTagName("input")[0].checked = false;
+        } else {
+            item.show()
+            item.elm.getElementsByTagName("input")[0].checked = false;
+            item.hide()
+        }
+        _id.push(item.values()["mols2grid-id"]);
+    });
     del_selection({{ grid_id | tojson }}, _id);
 });
 // invert
 $('#btn-chkbox-invert').click(function (e) {
-    var current_page = parseInt($("li.page-item.active > a").text());
-    var n_items = {{ n_items_per_page }};
-    var last_page = parseInt($("li.page-item > a").last().text());
     var _id_add = [];
     var _id_del = [];
     var _smiles = [];
-    for (let i = 0; i < last_page; i++) {
-        listObj.show(i * n_items + 1, n_items);
-        $("input:checkbox").each(function() {
-            this.checked = !this.checked;
-            var _id = parseInt($(this).closest(".cell").attr("data-mols2grid-id"));
-            if (this.checked) {
-                _id_add.push(_id);
-                _smiles.push($(this).siblings(".data-{{ smiles_col }}").first().text());
-            } else {
-                _id_del.push(_id);
-            }
-        });
-    }
-    listObj.show((current_page - 1) * n_items + 1, n_items);
+    listObj.items.forEach(function (item) {
+        if (item.elm) {
+            var chkbox = item.elm.getElementsByTagName("input")[0]
+            chkbox.checked = !chkbox.checked;
+        } else {
+            item.show()
+            var chkbox = item.elm.getElementsByTagName("input")[0]
+            chkbox.checked = !chkbox.checked;
+            item.hide()
+        }
+        if (chkbox.checked) {
+            _id_add.push(item.values()["mols2grid-id"]);
+            _smiles.push(item.values()["data-{{ smiles_col }}"]);
+        } else {
+            _id_del.push(item.values()["mols2grid-id"]);
+        }
+    });
     del_selection({{ grid_id | tojson }}, _id_del);
     add_selection({{ grid_id | tojson }}, _id_add, _smiles);
 });